Updated Controller
