+++ src/main/java/com/greenfox/lvlup/exception/DefaultExceptionHandler.java
@@ -1,14 +1,9 @@
 package com.greenfox.lvlup.exception;
 
 import com.greenfox.lvlup.model.CustomException;
-import com.greenfox.lvlup.model.ValidationError;
-import com.greenfox.lvlup.service.ValidationErrorBuilder;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.MethodArgumentNotValidException;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.ResponseBody;
-import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
 import org.springframework.web.servlet.mvc.method.annotation.ResponseEntityExceptionHandler;
 
@@ -21,5 +16,4 @@ public class DefaultExceptionHandler extends ResponseEntityExceptionHandler {
   public ResponseEntity<?> handleAllExceptions(CustomException e){
     return new ResponseEntity<>(e.getErrorMessage(), e.getHttpStatus());
   }
-
 }
